+++ CHANGELOG.rst
@@ -32,6 +32,7 @@ Added
     - Automated rebuilding of expired sheet configs (#904)
     - Material name suffix config and automated filling (#912)
     - ``get_node_obj()`` helper (#922)
+    - Update sheet config default protocols on sheet restore (#901)
 
 Changed
 -------